backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
nltk==3.8.1
speechrecognition==3.10.0
pyaudio==0.2.11
sqlite3
joblib==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
